Prevent error caused by conflicting manifests.
The MSW headers in the core project aren't marked as headers explicitly.  This means that non-.h files, such as the manifest files, are compiled into the project.  However, the manifest files cannot be compiled into the same project because they are for different architectures.  Therefore the build breaks.

This fixes the issue by marking them explicitly, in the same way as done in the base project.