Reverted to old behaviour for zlib, png, ... system libs
Commit by Dmitry would trigger an error when the system lib was not
found, instead of using the built-in library. The commit restores the
previous behaviour. Problem is now different: if you fill-in the values
for e.g. ZLIB (therefore ZLIB_FOUND becomes true) WXBUILD_SYSTEM_ZLIB
doesn't get toggled to ON (which is anyway debatable behaviour)